Hotfix/1.6.2: Fix bug where submitting transaction array thru v2 sendRawTransaction is impossible (#171)
Previously, sending a built transaction group to sendRawTransaction(signedTransactionGroup).do() would cause an err ReferenceError: txns is not defined. This was due to referencing a variable by the wrong name, txns, when it is stx_or_stxs. This hotfix repairs this reference error.

Review: https://github.com/algorand/js-algorand-sdk/pull/171