Add helpers for converting buffers to Transactions and vice versa. (#218)
It is a very common use case that js user-devs wish to take a Transaction, convert it to bytes, and then convert it back to a Transaction - either for reading and writing to file, coming in as bytes from a wire, or something else. The js-sdk does support this, but usage is non-obvious and error-prone. So, this commit adds three functions to algosdk: encodeUnsignedTransaction to turn a Transaction into bytes, decodeUnsignedTransaction to take those bytes back to a Transaction, and decodeSignedTransaction to take the output of Transaction.signTxn and turn it into something more usable. (decodeSignedTransaction would also work on the blob member of the object output by algosdk.signTransaction).

Review: https://github.com/algorand/js-algorand-sdk/pull/218
Resolves: https://github.com/algorand/js-algorand-sdk/issues/114 https://github.com/algorand/js-algorand-sdk/issues/176