Enabling a custom client for Algod and Indexer
Currently the only way for a dApp to use the SDK with an API service is to know the URL and token of the API service.
This prevents dApps from using API services / algod / indexer provided by the user or by a wallet such as AlgoSigner.

In particular:
* When using a wallet such as AlgoSigner, a dApp either need to manually serialize data and write path for endpoints (https://github.com/PureStake/algosigner/blob/develop/docs/dApp-integration.md#algosigneralgod-ledger-mainnettestnet-path-algod-v2-path--) or need to not use at all the provided API service.
* A user cannot use their preferred (potentially more trusted) API service / algod / index / API services
* A dApp cannot easily connect to a private network of the user

(See also https://github.com/algorandfoundation/ARCs/pull/8 for more details on wallets and discussion around them.)

The goal of this PR is to allow the above scenarios.
It does so by introducing an intermediate interface `BaseHTTPClient` and to allow to construct `Algodv2Client` and `IndexerClient` from such a `BaseHTTPClient`.

A `BaseHTTPClient` is minimal: it does not perform any serialization/de-serialization, does not fix `accept` headers, does not even try to utf8-decode strings. Its only purpose is to execute directly REST queries and return the raw responses.

`BaseHTTPClient`  is meant to be stable over releases of algosdk.