Added CRA polyfill solution text to save dev time
Added guide content to overcome the polyfill issue of Webpack 5 and above in react implementations, especially CRA-based ones. The solution, allows the fix for this problem without any patching or ejecting and opens the door to implement more custom Webpack plugins in action, just by a config line in the 'config-overrides.js' file.